Move `/dev/routes` API to `route-config` package
