+++ integration/vite-route-config-test.ts
@@ -47,7 +47,7 @@ test.describe("route config", () => {
           })]
         }
       `,
-      "app/routes.ts": `export default INVALID(`,
+      "app/routes.ts": `export const routes = [];`,
     });
     let buildResult = viteBuild({ cwd });
     expect(buildResult.status).toBe(1);
@@ -70,7 +70,7 @@ test.describe("route config", () => {
           })]
         }
       `,
-      "app/routes.ts": `export default INVALID(`,
+      "app/routes.ts": `export const routes = [];`,
     });
     let devError: Error | undefined;
     try {
@@ -121,13 +121,10 @@ test.describe("route config", () => {
     let files: Files = async ({ port }) => ({
       "vite.config.js": await viteConfig.basic({ port }),
       "app/routes.ts": js`
-        import { type RouteConfig } from "@react-router/dev/routes";
+        import { type RouteConfig, index } from "@remix-run/route-config";
 
         export const routes: RouteConfig = [
-          {
-            file: "test-route-1.tsx",
-            index: true,
-          },
+          index("test-route-1.tsx"),
         ];
       `,
       "app/test-route-1.tsx": `
@@ -185,13 +182,10 @@ test.describe("route config", () => {
         export { routes } from "./actual-routes";
       `,
       "app/actual-routes.ts": js`
-        import { type RouteConfig } from "@react-router/dev/routes";
+        import { type RouteConfig, index } from "@remix-run/route-config";
 
         export const routes: RouteConfig = [
-          {
-            file: "test-route-1.tsx",
-            index: true,
-          },
+          index("test-route-1.tsx"),
         ];
       `,
       "app/test-route-1.tsx": `
@@ -246,13 +240,10 @@ test.describe("route config", () => {
     let files: Files = async ({ port }) => ({
       "vite.config.js": await viteConfig.basic({ port }),
       "app/routes.ts": js`
-        import { type RouteConfig } from "@react-router/dev/routes";
+        import { type RouteConfig, index } from "@remix-run/route-config";
 
         export const routes: RouteConfig = [
-          {
-            file: "test-route-1.tsx",
-            index: true,
-          },
+          index("test-route-1.tsx"),
         ];
       `,
       "app/test-route-1.tsx": `
@@ -325,13 +316,10 @@ test.describe("route config", () => {
       "vite.config.js": await viteConfig.basic({ port }),
       "app/routes.ts": js`
         import path from "node:path";
-        import { type RouteConfig } from "@react-router/dev/routes";
+        import { type RouteConfig, index } from "@remix-run/route-config";
 
         export const routes: RouteConfig = [
-          {
-            file: path.resolve(import.meta.dirname, "test-route.tsx"),
-            index: true,
-          },
+          index(path.resolve(import.meta.dirname, "test-route.tsx")),
         ];
       `,
       "app/test-route.tsx": `
