--- packages/remix-dev/routes.ts
@@ -1,10 +1,0 @@
-export type { RouteConfig, RouteConfigEntry } from "./config/routes";
-
-export {
-  route,
-  index,
-  layout,
-  prefix,
-  relative,
-  getAppDirectory,
-} from "./config/routes";
