+++ packages/remix-route-config/index.ts
@@ -1,0 +1,13 @@
+export type {
+  UNSAFE_RouteConfig as RouteConfig,
+  UNSAFE_RouteConfigEntry as RouteConfigEntry,
+} from "@remix-run/dev";
+
+export {
+  route,
+  index,
+  layout,
+  prefix,
+  relative,
+  getAppDirectory,
+} from "./routes";
