+++ packages/remix-route-config/jest.config.js
@@ -1,0 +1,6 @@
+/** @type {import('jest').Config} */
+module.exports = {
+  ...require("../../jest/jest.config.shared"),
+  displayName: "route-config",
+  setupFiles: [],
+};
