+++ pnpm-workspace.yaml
@@ -18,6 +18,7 @@ packages:
   - "packages/remix-express"
   - "packages/remix-node"
   - "packages/remix-react"
+  - "packages/remix-route-config"
   - "packages/remix-serve"
   - "packages/remix-server-runtime"
   - "packages/remix-testing"
