+++ scripts/publish.js
@@ -62,6 +62,7 @@ async function run() {
     "serve",
     "css-bundle",
     "testing",
+    "route-config",
   ]) {
     publish(path.join(buildDir, "@remix-run", name), tag);
   }
